+++ CHANGELOG.md
@@ -2,6 +2,12 @@
 
 ## [Unreleased]
 
+### Added
+
+- [TD-1703] Include system and ancestry in data structure and data structure version show responses
+
+### Fixed
+
 - [TD-1747] Structures upload is not creating relation between structures and fields when including version number
 - [TD-1758] Structures with 'field' class are no indexed on create/upload